Changed the package name from services to handlers, getHttpResponse and handleHttpResponse handle exceptions within methods and use a logger, ExchangeRateService removed
